A GUI for supermodel. Not tested makefile but *should* work. To use it, just run supermodel with no command line paramaters. It's designed this way so people can still use old loaders / front ends if they want.

+++ Src/OSD/Windows/WinOutputs.cpp
@@ -241,7 +241,7 @@ LRESULT CWinOutputs::SendIdString(HWND hwnd, LPARAM id)
 		name = MapIdToName(id) ? MapIdToName(id) : "";
 
 	// Allocate memory for message
-	int dataLen = sizeof(CopyDataIdString) + name.length();
+	DWORD dataLen = sizeof(CopyDataIdString) + (DWORD)name.length();
 	CopyDataIdString *data = (CopyDataIdString*)new(nothrow) UINT8[dataLen];
 	if (!data)
 		return 1;
